Updated the game so that enemies can now move.
This is done using the moveEnemies method of the game class. This method calls each Enemy's doMove method, which moves the enemy to a location that is received from the path class. To get the next location, the path class is passed the current location of the unit, and it updates those values by doing some calculations depending on how the enemy is supposed to move.

The Enemy constructor now requires a path to be passed to it. 
The Path class has been updated with an abstract method, getNextLocation, which is used to get a point given the current location of the unit, or more specifically, an x and y coordinate. A diagonal path class has been created as a simple path which is not straight down movement.

There are some things that need to be done now:

Write tests for all the new and updated methods
Think about how to modify the speed of movement of units along their path. I suggest some kind of speed modifier in the path class, which can be modified using a setter method. The value given will then be used as a multiplier for the base values of the movement.

git-svn-id: https://codex.cs.bham.ac.uk/svn/teamjava/B3@118 88c3a9b8-8305-4942-bc19-8954e2f916ea