Created a path to be used by projectiles moving in a straight line.
Don't really like the way the directions are currently implemented - it would make sense to have these as some kind of inherited variable or something.

I think that probably, every path we implement will need some kind of directionality, so it would be good to have this as an enum in the main path class that all subclasses can access and use - we need at least some method of determining whether a projectile is going up or down, and for things that are going to be going diagonally, some way of telling which direction they're going in.

git-svn-id: https://codex.cs.bham.ac.uk/svn/teamjava/B3@120 88c3a9b8-8305-4942-bc19-8954e2f916ea